[Chore #162680798] Restructure helper functions to reduce duplication of files
