+++ app/api/v2/views/admin.py
@@ -4,8 +4,8 @@ from flask_restful import Resource
 # Local imports
 from app.api.helpers.auth_validation import (auth_required, access_control,
                                              auth_error_messages)
-from .errors import parser, get_error, Validation, error_messages
-from .helpers.incident_validation import (validate_admin_put_input)
+from app.api.helpers.errors import parser, get_error, Validation, error_messages
+from app.api.helpers.incident_validation import (validate_admin_put_input)
 from app.api.v2.models.incident import IncidentModel
 IncidentDB = IncidentModel()
 validator = Validation()
